fix: upload image in progress

+++ lib/src/domain/bloc/file_cubit/file_state.dart
@@ -9,7 +9,7 @@ class FileInitState extends FileState {}
 
 @freezed
 class UploadFileState extends FileState {
-  const factory UploadFileState.loading(int progress) = UploadingFileState;
+  const factory UploadFileState.loading(double progress) = UploadingFileState;
 
   const factory UploadFileState.success() = SuccessUploadingFileState;
